Only proceed to the next step when the context is established.
Before the code would always go to the next step, even if the GSS-API
library returned CONTINUE_NEEDED and the context not yet established.
The reason this problem does not appear to be exploitable is because a
Kerberos V5 GSS-API mechanism implementation would not return
CONTINUE_NEEDED and at the same time be able to successfully perform
gss_wrap and gss_unwrap.  Reported by Andreas Oberritter
<obi@saftware.de>.